Account Creation
There are errors with creating account in aSmack. For some reason, it
keeps throwing an exception every time. I uploaded the Smack version of
account creation as a basis and guideline for the code. Everything
should work perfectly and compile. I will use the skeleton code from
the the android XMPP demo to work with the code in aSmack.